requirements.txt
asgiref==3.4.1
Django==3.2.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
Pillow==8.3.1
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
